+++ packages/Bio/src/calculations/monomerLevelMols.ts
@@ -4,17 +4,17 @@ import * as DG from 'datagrok-api/dg';
 
 import wu from 'wu';
 
-import {getHelmMonomers} from '../package';
-import {SeqHandler} from '@datagrok-libraries/bio/src/utils/seq-handler';
 import {ISeqSplitted} from '@datagrok-libraries/bio/src/utils/macromolecule/types';
 import {GAP_SYMBOL} from '@datagrok-libraries/bio/src/utils/macromolecule/consts';
+
+import {_package, getHelmMonomers} from '../package';
 
 const V2000_ATOM_NAME_POS = 31;
 
 export async function getMonomericMols(
   mcol: DG.Column<string>, pattern: boolean = false, monomersDict?: Map<string, string>
 ): Promise<DG.Column> {
-  const sh = SeqHandler.forColumn(mcol);
+  const sh = _package.seqHelper.getSeqHandler(mcol);
   let molV3000Array;
   monomersDict ??= new Map();
   const monomers = sh.isHelm() ?