+++ packages/Bio/src/utils/helm-to-molfile/utils.ts
@@ -6,14 +6,14 @@ import {IMonomerLibBase} from '@datagrok-libraries/bio/src/types';
 
 import {SeqHelper} from '../seq-helper';
 
-import {_package, getMonomerLib} from '../../package';
+import {_package} from '../../package';
 
 /** Translate HELM column into molfile column and append to the dataframe */
 export async function getMolColumnFromHelm(
   df: DG.DataFrame, helmCol: DG.Column<string>, chiralityEngine: boolean = true, monomerLib: IMonomerLibBase
 ): Promise<DG.Column<string>> {
-  const seqHelper = await SeqHelper.getInstance();
-  const converter = seqHelper.getHelmToMolfileConverter(monomerLib);
+  const seqHelper: SeqHelper = _package.seqHelper as SeqHelper;
+  const converter = await seqHelper.getHelmToMolfileConverter(monomerLib);
   const molCol = converter.convertToRdKitBeautifiedMolfileColumn(helmCol, chiralityEngine, _package.rdKitModule, monomerLib);
   molCol.semType = DG.SEMTYPE.MOLECULE;
   return molCol;
@@ -22,9 +22,8 @@ export async function getMolColumnFromHelm(
 export async function getSmilesColumnFromHelm(
   helmCol: DG.Column<string>
 ): Promise<DG.Column<string>> {
-  const seqHelper = await SeqHelper.getInstance();
-  const monomerLib = getMonomerLib();
-  const converter = seqHelper.getHelmToMolfileConverter(monomerLib);
+  const seqHelper: SeqHelper = _package.seqHelper as SeqHelper;
+  const converter = await seqHelper.getHelmToMolfileConverter(_package.monomerLib);
   const smilesCol = converter.convertToSmiles(helmCol);
   smilesCol.semType = DG.SEMTYPE.MOLECULE;
   return smilesCol;