+++ packages/Bio/src/utils/macromolecule-column-widget.ts
@@ -3,7 +3,7 @@ import * as ui from 'datagrok-api/ui';
 import * as DG from 'datagrok-api/dg';
 
 import {TAGS as wlTAGS} from '@datagrok-libraries/bio/src/viewers/web-logo';
-import {SeqHandler} from '@datagrok-libraries/bio/src/utils/seq-handler';
+import {ISeqHelper} from '@datagrok-libraries/bio/src/utils/seq-helper';
 
 import {WebLogoViewer} from '../viewers/web-logo-viewer';
 
@@ -13,18 +13,17 @@ import {_package} from '../package';
 export class MacromoleculeColumnWidget extends DG.Widget {
   private viewed: boolean = false;
 
-  private readonly seqCol: DG.Column<string>;
-
   private wlViewer: WebLogoViewer | null = null;
 
-  constructor(seqCol: DG.Column<string>) {
+  constructor(
+    private readonly seqCol: DG.Column<string>,
+    private readonly seqHelper: ISeqHelper,
+  ) {
     super(ui.divV([]));
-
-    this.seqCol = seqCol;
   }
 
   async init(): Promise<void> {
-    const sh = SeqHandler.forColumn(this.seqCol);
+    const sh = this.seqHelper.getSeqHandler(this.seqCol);
     const pkgTooltipWebLogo = _package.properties.tooltipWebLogo;
     const colTooltipWebLogo = this.seqCol.getTag(wlTAGS.tooltipWebLogo);