+++ packages/Bio/src/utils/ui-utils.ts
@@ -1,11 +1,11 @@
 import * as grok from 'datagrok-api/grok';
 import * as DG from 'datagrok-api/dg';
 
-export function getMacromoleculeColumns(): DG.Column<any>[] | any {
+export function getMacromoleculeColumns(): DG.Column<string>[] {
   const columns = grok.shell.t.columns.bySemTypeAll(DG.SEMTYPE.MACROMOLECULE);
   if (columns === null) {
     grok.shell.error('Current table does not contain macromolecules');
-    return;
+    return [];
   }
   return columns;
 }